backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic[email]==2.9.2
python-multipart==0.0.24
python-docx==1.2.0
pdfplumber==0.11.5
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
aiosqlite==0.21.0
pydantic-settings==2.7.0
httpx==0.27.0
google-genai>=0.8.0
pypdf==6.10.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
